Make correction UI work with TPL imagery
