+++ oldto-site/corrections/index.html
@@ -56,7 +56,7 @@
     </a>
 
     <div class="details">
-      <div class="caption1"><a target="_blank" class="link" href="">City of Toronto Archives</a></div>
+      <div class="caption1"><a target="_blank" class="link source" href="">City of Toronto Archives</a></div>
       <div class="title"></div>
 
       <div class="value scope"></div>
@@ -64,14 +64,25 @@
       <div class="label">Date of Creation</div>
       <div class="value date"></div>
 
+      <!-- Toronto Archives-->
       <div class="label physical_desc">Physical Description</div>
       <div class="value physical_desc"></div>
 
-      <div class="label">Archival Citation</div>
+      <div class="label citation">Archival Citation</div>
       <div class="value"><a target="_blank" class="link citation" href=""></a></div>
 
-      <div class="label">Access conditions</div>
+      <div class="label condition">Access conditions</div>
       <div class="value condition"></div>
+
+      <!-- Toronto Public Library fields-->
+      <div class="label description">Notes</div>
+      <div class="value description"></div>
+
+      <div class="label creator">Creator</div>
+      <div class="value creator"></div>
+
+      <div class="label subject">Subject</div>
+      <div class="value subject"></div>
 
       <div style="display:none" class="geocode-debug"></div>
     </div>
