+++ oldto-site/index.html
@@ -158,7 +158,7 @@
     </div>
     <img class="inline-image"> <!-- note: this is only used on mobile -->
     <div class="details">
-      <div class="caption1"><a target="_blank" class="link" href="">City of Toronto Archives</a></div>
+      <div class="caption1"><a target="_blank" class="link source" href="">City of Toronto Archives</a></div>
       <div class="title"></div>
 
       <div class="value scope"></div>
@@ -185,6 +185,8 @@
 
       <div class="label subject">Subject</div>
       <div class="value subject"></div>
+
+      <!-- NB: keep in sync with corrections/index.html -->
 
       <div style="display:none" class="geocode-debug"></div>
